RS prototype choiceEval
This version evaluates whether the choice was correct or incorrect
based on the correctCue and “dispenses” water as appropriately. Note: I
have not actually written code to dispense water yet. Next step I need
to have a function that determines % correct and changes the cue (i.e.,
light to sound) after N trials and criterion is met, and then finally
terminates the experiment.